Moved V2 worker to it's folder
