+++ src/v2/ParsedMap.ts
@@ -1,5 +1,5 @@
 import {IInterval, IntervalHandler} from "../utils/IntervalHelper";
-import {IExportedValue, ObjectType} from "../worker/WorkerShared";
+import {IExportedValue, ObjectType} from "./worker/WorkerShared";
 
 interface IParsedTreeInterval extends IInterval {
     exp: IExportedValue;
